Creating utils file for manifesat handling
